requirements.txt
matrix-nio>=0.8.0
Markdown>=3.1.1
PyYAML>=5.1.2
Pillow
python-magic
python-olm
cachetools
atomicwrites
peewee
feedparser
giphypop
requests
